+++ internal/controller/k8s/k8scluster/cluster.go
@@ -18,10 +18,13 @@ package k8scluster
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
+	"strings"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/pkg/errors"
+	v1 "k8s.io/client-go/tools/clientcmd/api/v1"
 	"k8s.io/client-go/util/workqueue"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -136,19 +139,36 @@ func (c *externalCluster) Observe(ctx context.Context, mg resource.Managed) (man
 	cr.Status.AtProvider.ClusterID = meta.GetExternalName(cr)
 	cr.Status.AtProvider.State = clients.GetCoreResourceState(&observed)
 	clients.UpdateCondition(cr, cr.Status.AtProvider.State)
-
-	if kubeconfig, _, err = c.service.GetKubeConfig(ctx, cr.Status.AtProvider.ClusterID); err != nil {
-		c.log.Info(fmt.Sprintf("failed to get connection details. error: %v", err))
-	}
-
-	return managed.ExternalObservation{
+	mo := managed.ExternalObservation{
 		ResourceExists:          true,
 		ResourceUpToDate:        k8scluster.IsK8sClusterUpToDate(cr, observed),
 		ResourceLateInitialized: !cmp.Equal(current, &cr.Spec.ForProvider),
-		ConnectionDetails: managed.ConnectionDetails{
-			"kubeconfig": []byte(kubeconfig),
-		},
-	}, nil
+	}
+	if !strings.EqualFold(cr.Status.AtProvider.State, k8s.DEPLOYING) {
+		if kubeconfig, _, err = c.service.GetKubeConfig(ctx, cr.Status.AtProvider.ClusterID); err != nil {
+			c.log.Info(fmt.Sprintf("failed to get connection details. error: %v", err))
+		}
+	}
+	mo.ConnectionDetails = createKubernetesConnectionDetails(c, kubeconfig, mg)
+
+	return mo, nil
+}
+
+func createKubernetesConnectionDetails(c *externalCluster, kubeconfig string, mg resource.Managed) map[string][]byte {
+	var connectionConfig = map[string][]byte{
+		"kubeconfig": []byte(kubeconfig),
+	}
+
+	var clientkubeconfig v1.Config
+	if err := json.Unmarshal([]byte(kubeconfig), &clientkubeconfig); err != nil {
+		c.log.Info(fmt.Sprintf("failed to unmarshal connection details. error: %v", err))
+	} else {
+		connectionConfig["server"] = []byte(clientkubeconfig.Clusters[0].Cluster.Server)
+		connectionConfig["caData"] = clientkubeconfig.Clusters[0].Cluster.CertificateAuthorityData
+		connectionConfig["name"] = []byte(mg.GetName())
+		connectionConfig["token"] = []byte(clientkubeconfig.AuthInfos[0].AuthInfo.Token)
+	}
+	return connectionConfig
 }
 
 func (c *externalCluster) Create(ctx context.Context, mg resource.Managed) (managed.ExternalCreation, error) {